requirements.txt
requests~=2.32.5
loguru
python-dotenv